Book SQL data is working fine on the code
