+++ inc/config.inc.php
@@ -3,4 +3,4 @@
 define("DB_USER","root");
 define("DB_PASS","");
 define("DB_HOST","localhost");
-define("DB_NAME","wd5");
+define("DB_NAME","bookstest");
